Add benchmark for large PODs with properties
This benchmark should help diagnose #58.

On my machine with GCC 9.3, compilation for LargePod
with 100 properties used 3.4 GiB RAM and ran in 48s.

Complation for LargePod with 200 properties (PROPERTIES_2X)
used all of the system memory available to the VM (8 GiB),
and terminated after 5m30s.

This points to horrendeous scalability issues with
get_reader/has_writer, likely caused by the theoretically
quadratic factor blowup of the compilation of ProcessClass().
(Would have been quadratic if this was executed at runtime,
not sure how to properly calculate and describe it in this case).